Fill the normalized sample AudioBuffer in the appsink callbacks (#19) This allows taking advantage of the viualization queue thread to perform all the required computation as soon as the samples are received. There is no need to send the gst::Buffers through the channel: the AudioBuffer is now shared between the AudioController and the appsink.
Using a similar approach, the video widget box is now passed at Context creation. The video sink is constructed before the pipeline creation which ensures it is created in the GTK thread and which allows adding the widget to the video widget box without using channels.

Since channels are no longer required for realtime communication, the listener can be invoked less frequently.

This also fixes:
- Handling of media with multiple audio or video streams.
- Waveform flickering with some files (#18)