Use ControllerState to determine if a refresh must be triggered (#34) It was also an opportunity to move the requested_duration field to Rc<RefCell>> as the function connect_size_allocate was actually available in WidgetExt.
There is still an issue when zooming out to a certain point. The samples are not drawn on a buffer and the image seems to grow until it slows done the UI. This should be addressed as part of (#38). One lead to try is an enhancement of WaveformImage::render which would draw additional samples both before and after current image range.