Fix last flickerings on some samples (#64)
Last remaining variation between the iterative sample rendering and the link between samples chunk was due to the sequence by which the link was added. For `append_right`, the link from previous last sample to first added samples was performed after the new added samples were rendered. This didn't match the iterative sequence which would add this link before adding the new samples.

Note that there is no way to fix a similar effect for `append_left` as the link will be rendered above the first previously rendered samples. To solve this, a full redraw would be necessary. `append_right` is used in two cases:
1. When seeking backward in paused mode. In this case the flickering will not be disturbing as the application will present a single image until a user's action. When the next image is displayed, it is shifted so it shouldn't be easily noticeable. If the user resumes playback, the flickering might appear on a few samples. IMHO, it is acceptable.
2. In playback mode when seeking with the timeline to a position close enough to current first visible sample so that the available samples can be displayed and linked to the first visible sample. This case is rare enough to not introduce a dedicated detection that would engage a full redraw.