MediaInfo: use gst::TagList in place of a HashMap (#87)
The media module was designed in such a way that the underlining media library (first FFMpeg, then GStreamer) was hidden to the other modules. GStreamer is now all over the place and its tag handling is more approriate than using a HashMap. Exporting media with tags will be easier with gst::TagList and when the UI permits editing tags, it will be even more appropriate.