PlaybackContext: waveform pipeline without audioconvert (#100)
The `element` `audioconvert` can't be set in `async` mode. This prevents
the waveform sink from getting samples when the `pipeline` is in paused
state. This commit is an attempt to get rid of `audioconvert` so that the waveform
sink can pull buffers directly from the `queue.

Event though buffers are retrieved when the `pipeline` initialization is
complete, the waveform is not always displayed after the media is loaded.