Fix waveform not displaying after media is opened (#100)
Depending the initialization sequence, initialization of sample related
values in `WaveformBuffer` and `WaveformImage` caused the whole state of
this objects to reset, loosing data related to the display dimensions.

There is still an issue with the waveform being partially drawn in some
conditions. This can be reproduced frequently when switching chapters in
paused mode.