Waveform: fix clipped last position text (#100)
Last position text on waveform was clipped. This is due to a strange
behavior with the layout of the drawing area. The drawing area is
indirectly influenced with 2 columns of the primary grid of the UI.
Those 2 columns were added in order to allow the timeline to draw
while the play/pause & repeat buttons toolbar would keep its original
size. It appears that the clipping area of the context of the drawin
area is initially limitted to the length of the 2 columns of the
primary grid. The allocation of the drawing area is properly computed
and after the UI is updated a few times, the clipping area is also
properly computed.

Obvioulsy, there is a bug in GTK for this. In the meantime, the
solution is to move the chapter repeat button with the chapter toolbar.
This way, the 2 columns of the grid match the expected size for the
drawing area.