Add Two Steps seek (#103)
Two steps seek is used to get more context around current position in
Paused mode. After a seek in Paused mode, the cursor position was
centered if sufficient previous samples were available in the waveform
buffer. In some situation, e.g. after a chapter seek, this was not the
case and the cursor position would be the first sample to be displayed.

It was difficult for the user to step back a few sample in order to
adjust a chapter's boundary for instance. When the user wanted to search
backward and the first sample in the buffer was reached, the user couldn't
seek back anymore.

When necessary, this commit performs a first seek without syncrhonizing
the UI and when enough samples were received, a second seek is performed
to the initial target.

Note: there is still an issue when enlarging the window after e.g. a
chapter seek: the cursor disappears and the waveform is not rendered to
the correct position. This is probably due to the first_visible_sample in
the WaveformBuffer being initilized to the seek position.