SampleExtractor: use frame time instead of element clock (#102)
Remove the last systematic call to a gst locking function on each frame.
The time reference element was used to get the position for each frame.
Whether this element was used by querying its poistion or its clock was
used as a time reference, short hangs were perceptible in the waveform
scrolling. The reference element is the audio playback sink which is not
always immediately available since it is used during playback.

The widgets can access a frame clock with 1us accuracy. This is good enough
for the waveform scrolling and doesn't imply additional locks.