Add a CLI parameter to deactivate gtkglsink (#121)
Still need to check this on a system for which `gtkglsink` is operational.

`PlaybackContext` should handle the failure when the `video_sink` fails to
synchronize with its parent. This may occur on system where `gtkglsink` is
not operational and `gst-plugins-bad` is not installed. An error message
should be presented in such a case with an instruction to either install
`gst-plugins-bad` or to run the application with `--disable-gl`.