UI: fix wrong position after seeking (even more cases)
After, a new segment (e.g. after a seek), attempt to detect when audio playback
is getting a rate such as it is possible to use the frame presentation times
to estimate the actual audio position.

Use also this to determine when the first sample lock can be updated.

This is still not perfect, but it should lay the basis for subsequent
enhancements.