WaveformImage: reduce flickering effect
The waveform rendering incrementaly adds new samples to the existing
image. The rendering process links the newest samples to last sample
previously rendered.

There are 2 issues with this approach:

- The 2 `WaveformBuffer`s are not updated at the same time, which means that
the links appear at different ts for each buffer.
- In current code base, the rendering process draws all the samples in
current chunk for a single channel before moving to the next channel. This
prooved to be less CPU intensive as we can `stroke` several lines
intead of changing color at each line. However, this means that the links
are rendered above the already composed last samples which results in a
different composition compared to the regular samples (the ones that are not
part of link). Since the links appear at different places for the two
`WaveformBuffer`s, this produces a flickering effect as the `WaveformBuffer`s
alternate.

In this commit, we render each channel using an opaque color in order to
reduce the composition effect at links.