WaveformBuffer: use less sample overhead
The waveform rendering happens before the actual timestamp is known.
We use previous first sample as a reference and attempt to render
a full display range plus a few more samples to account for the
time spent between previous first sample and the actual first sample
when the waveform will be displayed. Since the rendering process builds
the whole waveform at each frame, it makes sens to optimize the overhead.
This commit reduces the overhead from a quarter of a display range to
an eighth. It doesn't reduce the CPU usage significantly.

Note that the actual overhead should be determine from the sample rate
and the frame rate.