Don't rely on dynamique dispatch for SampleExtractor
The only `SampleExtractor` implementer is `WaveformRenderer` and this
is kind of hard coded as it is selected when the `AudioController` is
instantiated. Using the trait `SampleExtractor` is still necessary to
keep the `media` module independent from `renderers`. While getting
ride of the dynamic usage of `SampleExtractor`, we can avoid the
`downcast`s to `WaveformRenderer` in `AudioController`.

There is no significant performance improvement.