Split: temporary workaround for some tracks with a toc
Some individual tracks (seen with a flac file) contain the whole album
toc. When split, we don't want to apply this to the individual track.

In current design, the TocVisitor ignores track entries, so the toc from
the above example is not handled. But there might be valid tocs with
tracks.