renderers: rework seek handling
Renderer bin catches the segment resulting from a seek and initiates
a 2 stages seek sequence if the conditions are met.

This is suboptimal as the first seek as already reached the source.
This rework experiments with catching the initial seek event and
either forwarding it as is in the regular seek case or replacing it
with the stage 1 seek for 2 stages seeks mode.